AIM Facial Index: real pagination, not a hardcoded 3-page cycle
The mugshot-grid browser's page button and arrow-key handlers only ever
cycled through three hardcoded stops (START_MERC = 0/40/80), regardless
of actual roster size - anything past page 3 (120 mercs) was simply
unreachable. The page-number label was the same story: a fixed lookup
table of six literal strings ("Page 1/2" .. "Page 1/1") that only knew
how to express 1-3 page scenarios.

- START_MERC/END_MERC widened UINT8 -> UINT16 (page-start offsets past
  255 didn't fit; needed for rosters approaching NUM_PROFILES's 2048).
- Replaced the 0/40/80 cycle with real math: numPages =
  ceil(MAX_NUMBER_MERCS / 40), cycling 0..numPages-1 in both directions.
  Same helper used by the page button and both arrow keys.
- Replaced the six-string gszAimPages lookup table with a single
  "Page %d/%d"-style template per language, formatted at runtime -
  updated consistently across all 8 localizations (English, Dutch,
  French, Italian, German, Polish, Russian, Chinese), preserving each
  language's existing word/abbreviation for "page".
- Found and fixed a live latent bug while in this code: EnterAimFacialIndex()
  reused a UINT8 loop variable to initialize gAimProfiles[] up to
  MAX_NUMBER_MERCS - past 255 mercs that wraps back to 0 before reaching
  the limit, an actual infinite loop/hang. Split into its own properly
  widened loop variable.

Full JA2.exe build compiles and links clean.

+++ Laptop/AimFacialIndex.h
@@ -27,7 +27,7 @@ BOOLEAN RenderAimFacialIndex();
 
 BOOLEAN DisplayAimFIMugShot();
 
-extern UINT8 START_MERC;
+extern UINT16 START_MERC;	// pagination fix: was UINT8 - page-start offsets past 255 (page 7+ at 40/page) didn't fit
 
 #endif
